refactor(phase4): migrate account components to src/features/account
Move AtCoderUserValidationForm, UserAccountDeletionForm, and
WarningMessageOnDeletingAccount from src/lib/components/ to
src/features/account/components/{settings,delete}/, and update
import paths in +page.svelte accordingly.
