+++ src/routes/users/edit/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { Tabs, TabItem, Alert } from 'flowbite-svelte';
 
-  // import AtCoderUserValidationForm from '$lib/components/AtCoderUserValidationForm.svelte';
-  import UserAccountDeletionForm from '$lib/components/UserAccountDeletionForm.svelte';
+  // import AtCoderVerificationForm from '$features/account/components/settings/AtCoderVerificationForm.svelte';
+  import AccountDeletionForm from '$features/account/components/delete/AccountDeletionForm.svelte';
   import ContainerWrapper from '$lib/components/ContainerWrapper.svelte';
   import FormWrapper from '$lib/components/FormWrapper.svelte';
   import LabelWrapper from '$lib/components/LabelWrapper.svelte';
@@ -127,7 +127,7 @@
           <span class="text-lg">アカウント削除</span>
         {/snippet}
 
-        <UserAccountDeletionForm {username} />
+        <AccountDeletionForm {username} />
       </TabItem>
     {/if}
   </Tabs>
